refact csv test.
